rest api에 version 넣기. #268
기본 version을 v00001로 지정, v00001과 v000705 두가지 버전 생성.
controllerTown과 controllerTown을 상속한 controllerTown24h를 생성.
routes/v000705/routeGather.js는 routes/v000001/routeGather.js를 그대로 사용.
