+++ server/app.js
@@ -14,10 +14,6 @@ var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
 var mongoose = require('mongoose');
 
-var routes = require('./routes/index');
-var users = require('./routes/users');
-
-var townForecast = require('./routes/routeTownForecast');
 var controllerManager = require('./controllers/controllerManager');
 var controllerShortRss = require('./controllers/controllerShortRss');
 /*
@@ -61,10 +57,9 @@ app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/', routes);
-app.use('/users', users);
-app.use('/town', townForecast);
-app.use('/gather', require('./routes/routeGather'));
+app.use('/', require('./routes/v000001'));
+app.use('/v000001', require('./routes/v000001'));
+app.use('/v000705', require('./routes/v000705'));
 
 global.curString = ['t1h', 'rn1', 'sky', 'uuu', 'vvv', 'reh', 'pty', 'lgt', 'vec', 'wsd'];
 //global.shortString = ['pop', 'pty', 'r06', 'reh', 's06', 'sky', 't3h', 'tmn', 'tmx', 'uuu', 'vvv', 'wav', 'vec', 'wsd'];
