+++ server/controllers/controllerTown24h.js
@@ -1,0 +1,224 @@
+/**
+ * Created by aleckim on 2016. 2. 19..
+ * controllerTown에서 short의 time table을 0~21 에서 3~24시로 변경한 클래스임.
+ */
+
+var ControllerTown = require('../controllers/controllerTown');
+var kmaTimeLib = require('../lib/kmaTimeLib');
+
+/**
+ *
+ * @constructor
+ */
+function ControllerTown24h() {
+    //var self = this;
+
+    ControllerTown.call(this);
+
+    /**
+     * adjust tmn, tmx and reorganize position of tmn and tmx
+     * @param req
+     * @param res
+     * @param next
+     */
+    this.adjustShort = function(req, res, next) {
+        var regionName = req.params.region;
+        var cityName = req.params.city;
+        var townName = req.params.town;
+
+        var meta = {};
+        meta.method = 'adjustShort';
+        meta.region = regionName;
+        meta.city = cityName;
+        meta.town = townName;
+        log.info('>', meta);
+
+        if (!req.hasOwnProperty('short')) {
+            log.error("Short forecast data hasn't attached on req");
+            next();
+            return this;
+        }
+
+        var daySummaryList = [];
+        req.short.forEach(function (short, index) {
+
+            //client 하위 버전 지원 못함.
+            if (short.time === "0000") {
+                var D = kmaTimeLib.convertStringToDate(short.date);
+                D.setDate(D.getDate()-1);
+                //date = back one day
+                //date = (parseInt(short.date)-1).toString();
+                short.time = "2400";
+                short.date = kmaTimeLib.convertDateToYYYYMMDD(D);
+            }
+
+            var daySummary = self._createOrGetDaySummaryList(daySummaryList, short.date);
+            if (daySummary.taMax < short.t3h) {
+                daySummary.taMax = short.t3h;
+            }
+            if (daySummary.taMax < short.tmx) {
+                daySummary.taMax = short.tmx;
+                log.verbose(index+" tmx clear");
+                //clear tmx
+            }
+            short.tmx = -50;
+
+            if (daySummary.taMin === -50 && short.t3h !== -50) {
+                daySummary.taMin = short.t3h;
+            }
+            else if (daySummary.taMin > short.t3h && short.t3h !== -50) {
+                daySummary.taMin = short.t3h;
+            }
+
+            if (daySummary.taMin === -50 && short.tmn !== -50) {
+                daySummary.taMin = short.tmn;
+            }
+            else if (daySummary.taMin > short.tmn && short.tmn !== -50) {
+                daySummary.taMin = short.tmn;
+                short.tmn = -50;
+                log.verbose(index+" tmn clear");
+            }
+            short.tmn = -50;
+        });
+
+        req.short.forEach(function (short) {
+            var daySum = self._createOrGetDaySummaryList(daySummaryList, short.date);
+            if (daySum.taMax === -50 || daySum.taMin === -50) {
+                log.error("short date:"+short.date+" fail to get daySummary");
+                return;
+            }
+            if (short.time === "0600") {
+                short.tmn = daySum.taMin;
+            }
+            if (short.time === "1500") {
+                short.tmx = daySum.taMax;
+            }
+        });
+
+        //client 하위 버전 지원 못함.
+        req.short.forEach(function (short, index) {
+
+            var daySum = self._createOrGetDaySummaryList(daySummaryList, short.date);
+            if (daySum.taMax === -50 || daySum.taMin === -50) {
+                log.error("short date:"+short.date+" fail to get daySummary");
+                return;
+            }
+            var tmxDiff = Math.abs(daySum.taMax - short.t3h);
+            var tmnDiff = Math.abs(daySum.taMin - short.t3h);
+
+            if (!daySum.hasOwnProperty("tmxDiff")) {
+                daySum.tmxDiff = tmxDiff;
+                short.tmx = daySum.taMax;
+                daySum.tmxIndex = index;
+            }
+            else {
+                if (daySum.tmxDiff === tmxDiff) {
+                    if(short.time === "1500")  {
+                        req.short[daySum.tmxIndex].tmx = -50;
+
+                        daySum.tmxDiff = tmxDiff;
+                        short.tmx = daySum.taMax;
+                        daySum.tmxIndex = index;
+                        log.verbose("put index:"+index+" tmx:"+short.tmx);
+                    }
+                    else if (req.short[daySum.tmxIndex].time === "1500") {
+                        //skip
+                    }
+                    else {
+                        //nearest from 1500
+                        //late time
+                        req.short[daySum.tmxIndex].tmx = -50;
+
+                        daySum.tmxDiff = tmxDiff;
+                        short.tmx = daySum.taMax;
+                        daySum.tmxIndex = index;
+                        log.verbose("put index:"+index+" tmx:"+short.tmx);
+                    }
+                }
+                else if (daySum.tmxDiff > tmxDiff) {
+                    req.short[daySum.tmxIndex].tmx = -50;
+
+                    daySum.tmxDiff = tmxDiff;
+                    short.tmx = daySum.taMax;
+                    daySum.tmxIndex = index;
+                    log.verbose("put index:"+index+" tmx:"+short.tmx);
+                }
+            }
+
+            if (!daySum.hasOwnProperty("tmnDiff")) {
+                daySum.tmnDiff = tmnDiff;
+                short.tmn = daySum.taMin;
+                daySum.tmnIndex = index;
+            }
+            else {
+                if (daySum.tmnDiff === tmnDiff) {
+                    if(short.time === "0600")  {
+                        req.short[daySum.tmnIndex].tmn = -50;
+
+                        daySum.tmnDiff = tmnDiff;
+                        short.tmn = daySum.taMin;
+                        daySum.tmnIndex = index;
+                        log.verbose("put index:"+index+" tmn:"+short.tmn);
+                    }
+                    else if (req.short[daySum.tmnIndex].time === "0600") {
+                        //skip
+                    }
+                    else {
+                        //nearest from 0600
+                        //early time
+                    }
+                }
+                else if (daySum.tmnDiff > tmnDiff) {
+                    req.short[daySum.tmnIndex].tmn = -50;
+
+                    daySum.tmnDiff = tmnDiff;
+                    short.tmn = daySum.taMin;
+                    daySum.tmnIndex = index;
+                    log.verbose("put index:"+index+" tmn:"+short.tmn);
+                }
+            }
+        });
+
+        next();
+        return this;
+    };
+
+}
+
+// subclass extends superclass
+ControllerTown24h.prototype = Object.create(ControllerTown.prototype);
+ControllerTown24h.prototype.constructor = ControllerTown24h;
+
+/**
+ *
+ * the day before yesterday 03h ~ the day after tomorrow 24h
+ * @returns {Array}
+ * @private
+ */
+ControllerTown24h.prototype._makeBasicShortList = function() {
+    var result = [];
+    var self = this;
+
+    var currentTime = parseInt(self._getCurrentTimeValue(9).time.slice(0,2));
+
+    for(var i=0 ; i < 40 ; i++){
+        var item = self._getTimeValue(9-currentTime-24*2+(i*3)+3);
+        shortString.forEach(function(string){
+            if(string == 'tmn' || string === 'tmx' || string === 't3h') {
+                item[string] = -50;
+            } else if (string === 'uuu' || string === 'vvv') {
+                item[string] = -100;
+
+            }else{
+                item[string] = -1;
+            }
+        });
+        result.push(item);
+    }
+
+    self._dataListPrint(result, 'route S', 'template short');
+
+    return result;
+};
+
+module.exports = ControllerTown24h;
