+++ server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "TodayWeather",
-  "version": "0.0.1",
+  "version": "0.7.5",
   "description": "TodayWeather: Inform how warm or cold it is today than yesterday.",
   "repository": {
     "type": "git",
